refactor: divide up util functions
